+++ custom_components/spotifyplus/__init__.py
@@ -1830,7 +1830,7 @@ async def async_setup_entry(hass:HomeAssistant, entry:ConfigEntry) -> bool:
                 # - get a reference to the HA OAuth2 session instance.
                 # - call `async_refresh_token` to refresh the expired token.
                 # - set a token `status` attribute to denote the configuration update is for a token refresh event.  this allows the
-                #   `options_update_listener` to bypass the confgiuration reload processing since we only need to reload the configuration
+                #   `options_update_listener` to bypass the configuration reload processing since we only need to reload the configuration
                 #   if the user initiated an options change via the UI (e.g. no need to reload the configuration for token updates).
                 # - call the `add_job` method to add a job that will call `async_update_entry` to persist the refreshed token to config storage.
                 # - return the refreshed token to the caller.
@@ -2004,6 +2004,9 @@ async def async_unload_entry(hass:HomeAssistant, entry:ConfigEntry) -> bool:
     to be removed, as they are already removed by the time this method is called.
     This is accomplished by the "entry.async_on_unload(listener)" call in async_setup_entry,
     which removes them from the configuration entry just before it is unloaded.
+
+    Note that something changed with HA 2024.6 release that causes the `update_listeners` array 
+    to still contain entries; prior to this release, the `update_listeners` array was empty by this point.
     """
     try:
 
@@ -2025,8 +2028,10 @@ async def async_unload_entry(hass:HomeAssistant, entry:ConfigEntry) -> bool:
 
             # a quick check to make sure all update listeners were removed (see method doc notes above).
             if len(entry.update_listeners) > 0:
-                _logsi.LogArray(SILevel.Warning, "'%s': Component configuration update_listener(s) did not get removed before configuration unload (%d items - should be 0)" % (entry.title, len(entry.update_listeners)), entry.update_listeners)
-                entry.update_listeners.clear()
+                _logsi.LogArray(SILevel.Warning, "'%s': Component configuration update_listener(s) did not get removed before configuration unload (%d items - should be 0 prioer to HA 2026.0 release, but after that release still contains entries)" % (entry.title, len(entry.update_listeners)), entry.update_listeners)
+                # 2024/06/08 - I commented out the following line to clear the `update_listeners`, as it was causing `ValueError: list.remove(x): x not in list`
+                # exceptions starting with the HA 2024.6.0 release!
+                #entry.update_listeners.clear()
 
         # return status to caller.
         _logsi.LogVerbose("'%s': Component async_unload_entry completed" % entry.title)